customise Pod::Perldoc to fix output misbehaviour
Pod::Perldoc has, since version 3.20, exhibited various kinds of
misbehaviour relating to a bad default choice of formatter.  Output has
sometimes appeared mangled due to the newly-default formatter emitting
unportable escape sequences, and sometimes there has been a more
severe output failure due to perldoc making unportable changes to pager
configuration in an attempt to make the escape sequences work.  This is
discussed in [perl #131762].  In the upstream instance of the module
there have been tweaks to the unportable behaviour, but not an actual fix.

In order to make the core distro ship a reliably-working version of
perldoc, this patch customises Pod::Perldoc to implement the obvious
fix for the portability problems.  The fixed version defaults to the
ToText formatter, which produces properly plain text that will go through
any pager.  It never attempts any change to pager configuration.